Make the MDL state fully contained in the DeviceGlobalState
This will help moving toward using the internal visrtx code to handle samplers and other objects.
